seCODEcret
My little project's aim is to simulate a security app which encrypts any given text by a given number of shifts. For example, the users wants to encrypt "abc", shifting every letter with 2 positions. The app will generate the text "cde"(a+2=c, b+2=d, c+2=e). The use of this app would be helpful in private messaging.